requirements.streamlit.txt
streamlit==1.26.0
pandas==2.1.0
sqlalchemy==1.4.50
psycopg2-binary==2.9.7
plotly==5.16.1
mlflow==2.15.0
python-dotenv
boto3
xgboost
scikit-learn==1.5.2
openpyxl